Move component to ContactUsForm
